YKI(Frontend): Refactor registration components
