+++ frontend/packages/yki/src/components/registration/steps/register/EmailRegistrationDetails.tsx
@@ -1,31 +1,18 @@
-import OpenInNewIcon from '@mui/icons-material/OpenInNew';
-import {
-  Checkbox,
-  FormControlLabel,
-  Link,
-  Radio,
-  RadioGroup,
-} from '@mui/material';
+import { FormControlLabel, Radio, RadioGroup } from '@mui/material';
 import { ChangeEvent, useState } from 'react';
-import { CustomTextField, H2, Text } from 'shared/components';
-import { Color, TextFieldTypes } from 'shared/enums';
+import { CustomTextField, Text } from 'shared/components';
+import { TextFieldTypes } from 'shared/enums';
 import { InputFieldUtils } from 'shared/utils';
 
-import { PersonDetails } from 'components/registration/steps/PersonDetails';
+import { PersonDetails } from 'components/registration/steps/register/PersonDetails';
 import { useCommonTranslation, usePublicTranslation } from 'configs/i18n';
-import { useAppDispatch } from 'configs/redux';
-import { CertificateLanguage, RadioButtonValue } from 'enums/app';
-import {
-  PublicEmailRegistration,
-  RegistrationCheckboxDetails,
-} from 'interfaces/publicRegistration';
+import { useAppDispatch, useAppSelector } from 'configs/redux';
+import { RadioButtonValue } from 'enums/app';
+import { PublicEmailRegistration } from 'interfaces/publicRegistration';
 import { updatePublicRegistration } from 'redux/reducers/registration';
+import { registrationSelector } from 'redux/selectors/registration';
 
-export const EmailRegistration = ({
-  registration,
-}: {
-  registration: PublicEmailRegistration;
-}) => {
+export const EmailRegistrationDetails = () => {
   const { t } = usePublicTranslation({
     keyPrefix: 'yki.component.registration.registrationDetails',
   });
@@ -50,16 +37,8 @@ export const EmailRegistration = ({
   });
 
   const dispatch = useAppDispatch();
-
-  const handleCheckboxClick = (
-    fieldName: keyof RegistrationCheckboxDetails
-  ) => {
-    dispatch(
-      updatePublicRegistration({
-        [fieldName]: !registration[fieldName],
-      })
-    );
-  };
+  const registration: Partial<PublicEmailRegistration> =
+    useAppSelector(registrationSelector).registration;
 
   const getEventTargetValue = (value: string) => {
     if (value === (RadioButtonValue.YES as string)) {
@@ -125,14 +104,8 @@ export const EmailRegistration = ({
   });
 
   return (
-    <div className="margin-top-xxl rows gapped">
-      <H2>{t('title')}</H2>
-      <Text>
-        {t('description1')}
-        <br />
-        {t('description2')}
-      </Text>
-      <div className="email-registration-details rows gapped margin-top-sm">
+    <>
+      <div className="registration-details rows gapped margin-top-sm">
         <PersonDetails
           getCustomTextFieldAttributes={getCustomTextFieldAttributes}
         />
@@ -195,88 +168,6 @@ export const EmailRegistration = ({
           />
         )}
       </div>
-      <div>
-        <Text>
-          <b>{t('certificateLanguage')}</b>
-        </Text>
-        <RadioGroup row onChange={handleChange('certificateLanguage')}>
-          <FormControlLabel
-            className="radio-group-label"
-            value={CertificateLanguage.FI}
-            control={<Radio />}
-            label={translateCommon('languages.fin')}
-          />
-          <FormControlLabel
-            className="radio-group-label"
-            value={CertificateLanguage.SV}
-            control={<Radio />}
-            label={translateCommon('languages.swe')}
-          />
-          <FormControlLabel
-            className="radio-group-label"
-            value={CertificateLanguage.EN}
-            control={<Radio />}
-            label={translateCommon('languages.eng')}
-          />
-        </RadioGroup>
-      </div>
-
-      <H2>{t('termsAndConditions.title')}</H2>
-      <Text>
-        <b>{t('termsAndConditions.description1')}</b>
-        <br />
-        <b>{t('termsAndConditions.description2')}</b>
-        <br />
-        {t('termsAndConditions.description3')}
-      </Text>
-      <FormControlLabel
-        control={
-          <Checkbox
-            onClick={() => handleCheckboxClick('termsAndConditionsAgreed')}
-            color={Color.Secondary}
-            checked={registration.termsAndConditionsAgreed}
-          />
-        }
-        label={t('termsAndConditions.label')}
-        className="public-registration__grid__preview__privacy-statement-checkbox-label"
-      />
-      <div>
-        <Text>{t('privacyStatement.description')}:</Text>
-        <div className="columns gapped-xxs">
-          <Link href={translateCommon('privacyStatementLink')} target="_blank">
-            <Text>{t('privacyStatement.linkLabel')}</Text>
-          </Link>
-          <OpenInNewIcon />
-        </div>
-      </div>
-      <FormControlLabel
-        control={
-          <Checkbox
-            onClick={() => handleCheckboxClick('privacyStatementConfirmation')}
-            color={Color.Secondary}
-            checked={registration.privacyStatementConfirmation}
-          />
-        }
-        label={t('privacyStatement.label')}
-        className="public-registration__grid__preview__privacy-statement-checkbox-label"
-      />
-      <H2>{t('whatsNext.title')}</H2>
-      <Text>{t('whatsNext.description1')}</Text>
-      <Text>
-        {t('whatsNext.description2')}
-        <br />
-        {t('whatsNext.description3')}:
-        <br />
-        <div className="columns gapped-xxs">
-          <Link
-            href={translateCommon('specialArrangementsLink')}
-            target="_blank"
-          >
-            {t('whatsNext.linkLabel')}
-          </Link>
-          <OpenInNewIcon />
-        </div>
-      </Text>
-    </div>
+    </>
   );
 };
