+++ frontend/packages/yki/src/styles/components/registration/_public-registration.scss
@@ -14,7 +14,7 @@
         }
       }
 
-      .email-registration-details {
+      .registration-details {
         max-width: 720px;
 
         &__address-grid {
